autotest: WiFi: Allow stopping of individual hostapd instances
Allow the "deconfig" command to take an instance number to
be removed.  In order to allow this to happen, we need to
be able to do an individual interface remove (for "silent mode"
deconfig) and hostapd kill for a specific interface.

BUG=chrome-os-partner:15005
TEST=Tested for regression using network_WiFiRoaming/006BeaconLoss.
Testing for functionality using new PMKSA test.

Change-Id: Icb146b2759bf65caef931b65d81a2a90c7309234
Reviewed-on: https://gerrit.chromium.org/gerrit/39320
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>