Add full power_LoadTest run to power_weekly/power_requirements suite
power_LoadTest has only ever been run as part of the power_daily suite
in recent times. Since it can take a significant amount of time (up to
about 9 hours), that run was changed to a hard one hour limit in commit
623a149. However, it would still be useful to at least occasionally run
the full test (until the battery is drained) and get an accurate battery
life measurement.

This patch adds such a full power_LoadTest run to the
power_weekly/power_requirements suite which runs once a week. The
existing 1hour runs in power_daily remain unchanged. It also fixes the
power_BatteryCharge test to abort automatically when the battery is
full, allowing simple charging to full capacity without the need to
account for design capacity vs. battery wear.

BUG=chromium-os:34798
TEST=None... how do I test a suite? Is there a good way to do this?

Change-Id: I4e3f58277e86a013b8573808874db4154ac02894
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/39336
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>