firmware_TouchpadMTB: Fixed a bug in get_x_y_multiple_slots
The evdev device driver only posts changes of device attributes to
the device node. Hence, we need to keep the states of the attributes
such as SLOT, X, Y, PRESSURE, etc. when parsing the packets.

In the function get_x_y_multiple_slots(), an temporary iterating
variable 'slot' is accidentally used with the same name of 'slot'
which tracks the current slot. This CL fixed the bug and added
a unit test to ensure that it tracks the slot number in the
underlying state machine properly.

BUG=chromium-os:35024
TEST=Run the following unit tests on a host or a chromebook.
$ cd .../firmware_TouchpadMTB
$ python tests/mtb_unittest.py
All unit tests should pass.

Change-Id: Ia032f42464bba6f7be3c9bfea0b7136f237d763e
Reviewed-on: https://gerrit.chromium.org/gerrit/39448
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>