firmware_TouchpadMTB: Test the segments of middle and both ends for fat fingers
When drawing with the thumb edge or with fat fingers, it usually
generates some wobbles when the fingers are touching the pad. In this
CL, we adjust the test_conf to display the test results of
LinearityValidator in the middle segment and at both end segments
separately.

BUG=chromium-os:35752
TEST=Run the following command on a chromebook.
$ DISPLAY=:0 sudo python main.py
Observe the test results of LinearityMiddleValidator and
LinearityBothEndsValidator for the gestures:
  - drag_edge_thumb, and
  - two_fat_fingers_tracking

The test result should display both LinearityMiddleValidator and
LinearityBothEndsValidator.

Change-Id: I441ad8f83201c8adf89ec43d7413ba68525cf296
Reviewed-on: https://gerrit.chromium.org/gerrit/39450
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>