faft: Speed-up FwScreenPressPower and FwScreenCloseLid tests
This CL supports a timeout argument to confirm DUT shutdown. We limit
the shutdown check only on the dev screen internal disk boot. Since the
other scenarios, like recovery insert screen, are impossible to boot.
And we also reduce this timeout value to 60s (default 100s) which is
able to confirm DUT shutdown in the case of the internal disk boot.

This CL reduces the test runtime from 20min to 10min on Link.

BUG=chromium-os:34392
TEST=run FwScreenPressPower and FwScreenCloseLid passed on Link.

Change-Id: I4e8059e2e2ada29d3c3da52526656f2002a9823a
Reviewed-on: https://gerrit.chromium.org/gerrit/39583
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>