firmware_TouchpadMTB: set basic as the default type for robot line gestures
When determining "basic" or "swipe" for the robot line gestures, the
default type is set to "basic" except when the "swipe" string could
be found in the gesture name.

BUG=chromium-os:35751
TEST=Run the command on a chromebook with or without the robot.
$ DISPLAY=:0 OPTIONS="-m ROBOT_SIM -s" python main.py

Observe that in the 2nd gesture <one_finger_to_edge> it displays
something like:

  Executing: "python .../site-packages/gestures/line.py
  lumpy 0.500000 0.500000 -0.050000 0.500000 33.000000 basic"

It should display "basic" at the end of the script command options.

Change-Id: I1a8332a0573a9df73d38df03384bebba920c8e21
Reviewed-on: https://gerrit.chromium.org/gerrit/40206
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Tested-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>