autotest: WiFi: Create test for PMKSA caching
Verify that client can use PMKSA caching to roam between APs.
It significantly speeds up the association process if the client
can prove to the AP that it is in possession of cached credentials
for the network instead of going through the EAP-TLS process.

In order to create this test, add functions to command the DUT to
roam to an AP instance, as well as a command to verify using the
hostapd logs that PMKSA caching was employed for authentication.

BUG=chrome-os-partner:15005
TEST=Run test

Change-Id: Ia0f40957ea2d80ad1fb693497373d2fb38a5ca9c
Reviewed-on: https://gerrit.chromium.org/gerrit/39894
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>