Autotest: GS_Offloader - Multiple Threads offloading same Dir's
It appears that the new multiple thread gs_offloader code is offloading
the same directories multiple times. Looking at the code it appears
that the directories are being queued up multiple times.

In order to address this, we will lock a directory by touching a file it,
and if locked don't queue it up again.

BUG=chromium-os:37297
TEST=local setup, ensured the lockfile is created and these directories
     are not being processed multiple times.

Change-Id: I57b0bf53d816b0683fb967be561f32fb0673a95e
Reviewed-on: https://gerrit.chromium.org/gerrit/39800
Reviewed-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Simran Basi <sbasi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>