faft: Trim down the content of RW main firmware body
When we build the main firmware image, the signing utility uses the size of
the input file as the size of the data to sign. This input file is smaller
than the whole fmap section area.

However, in FAFT, we dump the data as large as the whole fmap section size.
So any resigning generates a different VBlock header, which creates an issue
on Depthcharge. So this CL is to trim down the content of RW main firmware
body, making it the same size of its original.

BUG=chrome-os-partner:16678
TEST=run RollbackFirmware and RONormalBoot, see the resulted VBLOCK as before

Change-Id: Ia1e59100f843ac3d53027293492f92850423f6e8
Reviewed-on: https://gerrit.chromium.org/gerrit/39830
Reviewed-by: Vic Yang <victoryang@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>