Autotest GSOffloader: Make email messages and log files consistent
Currently when we send out an email for gs_offloader issues, we only
send a snippet of the stderr message. This is due to the fact the
error message can be hundreds of lines long. And in the log we have
all the output from stdout and stderr saved.

For consistency sake and to make it easier for a lab sheriff, etc
who may see the emails to find the same message in the logs, now we
will log the shorter email message followed by the more dense logs.

BUG=chromium-os:37298
TEST=None

Change-Id: Ic46f475f00196e399428f79b518511d78455b9ac
Reviewed-on: https://gerrit.chromium.org/gerrit/39860
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Simran Basi <sbasi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>