Add "Hot CPU" suspend test
This adds a test that suspend the system while at 100% CPU usage.

I am not sure how useful this is, but we thought it might expose races
in the suspend code that would not happen on an idle system. The race
would likely need to happen before userspace is frozen.

TEST=run_remote_tests.sh power_HotCPUSuspend
BUG=none

Change-Id: Ic4251ff4c45ee8d7c88c973a910fb035f72236e2
Reviewed-on: https://gerrit.chromium.org/gerrit/40342
Reviewed-by: Michael Spang <spang@chromium.org>
Tested-by: Michael Spang <spang@chromium.org>
Commit-Queue: Michael Spang <spang@chromium.org>