Add power_SuspendStress autotest with different payloads
This patch adds the first version of a new power_SuspendStress autotest.
It provides the basic code to cause repeated suspend/resume cycles over
a given amount of time, and is intended to be run in parallel with
existing functional tests through different control files. This first
set contains control files for an OpenGL video test, a disk IO stressing
benchmark and a bare run of just the stress test itself, with a few
simple checks for general system status afterwards. All variants exist
in a BVT version targeted at a few minutes and a daily run that lasts
between one and two hours.

BUG=chromium-os:35901
TEST=Check to see if BVT and daily iterations of this start running in
the lab.

Change-Id: I6f555ec0cc7fa60ab7d96296fc573f669297146f
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/40071
Reviewed-by: Sameer Nanda <snanda@chromium.org>