[autotest] Add a barrier after |os.walk()|
We have too many issues with the threading of gs_offloader because now
the main loop of finding directories to offload is trying to process
directories that are currently offloading. If we add a barrier after
looping over all directories in the main thread, then we preserve the
original single-threaded semantics.

TEST=ran on chromeos-autotest.cbf
BUG=chromium-os:37453

Change-Id: I76f91408f300fc1d6fb6cb7d59223e871d815865
Reviewed-on: https://gerrit.chromium.org/gerrit/40318
Commit-Queue: Alex Miller <milleral@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Tested-by: Alex Miller <milleral@chromium.org>