[autotest] Bump num for kernel_daily_regression to 2.
The addition of DEPENDENCIES="servo" to

    server/site_tests/power_SuspendShutdown/control

made the test not able to be scheduled because there's now two different
HostSpecs and num=1 by default. This means that you require two machines
(one with and one without servo), but were only asking for one machine
(num=1) so the power_SuspendShutdown test was never getting run.

TEST=new suite_scheduler.py --sanity
BUG=None

Change-Id: I73f97c27738ba6cc0cb6610a2a8a5c5323b386ff
Reviewed-on: https://gerrit.chromium.org/gerrit/40414
Tested-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>