firmware_TouchpadMTB: fix a bug about division in firmware_summary
This CL fixed a bug by replacing

  average = float(numerator / denominator)       ... (1)

with

  average = float(numerator) / denominator       ... (2)

When both numerator and denominator are integers, the value derived
in formula (1) will be truncated before the float() is applied.
Hence, convert the numerator to its float value before the division
is performed. Unit tests are added for verification.

BUG=chromium-os:37494
TEST=Run the command on a host:
$ cd .../autotest/files/client/site_tests/firmware_TouchpadMTB
$ python tests/firmware_summary_unittest.py
All unit tests must pass.

Change-Id: I85c1614139ab06a7582cec1fbaf55594b051211e
Reviewed-on: https://gerrit.chromium.org/gerrit/40423
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>