firmware_TouchpadMTB: add a manual mode
This CL adds a manual mode for recording gestures. While the complete
mode includes all gestures, the manual mode includes the gestures
in the complete mode minus the gestures in the robot mode.

This mode is useful when a tester wants to record the gestures not
performed by the robot.

Hence, the process of testing a firmware drop would be

Step 1: Records gestures for a firmware drop
Step 1.1: Use the robot to record the gestures in the robot mode.
          Typically, at least 3 rounds are recommended.
          $ DISPLAY=:0 OPTIONS="-m ROBOT" python main.py
Step 1.2: A tester records the gestures in the manual mode.
          Repeat as many rounds as in step 1.
          $ DISPLAY=:0 OPTIONS="-m MANUAL" python main.py
Step 2: repeat step 1 for the other firmware drop.
Step 3: Derive the summary of the test results.
Step 3.1: copy all logs to a tmp directory, say /tmp/logs.
          Note that $LOG_DIR are the directories generated above.
          $ cp /var/tmp/touchpad_firmware_test/$LOG_DIR/*.log /tmp/logs
Step 3.2: Derive the summary.
          $ python firmware_summary.py /tmp/logs

BUG=chromium-os:37494
TEST=Run the command on a host:
$ cd .../autotest/files/client/site_tests/firmware_TouchpadMTB
$ DISPLAY=:0 OPTIONS="-m MANUAL" python main.py

Observe the following message: "Note: the MANUAL mode is used."
Also observe that only a subset of gestures will be recorded.

Change-Id: I86e3afe46bee39029af9be731e851aa5f4dbe06e
Reviewed-on: https://gerrit.chromium.org/gerrit/40428
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>