firmware_TouchpadMTB: add the firmware version and mode in the directory
Since some tests will be conducted using the robot, and some other
tests by human testers, it is clearer to add firmware version and
mode in the test directory and the log files.

BUG=chromium-os:37494
TEST=Run the command on a chromebook:
$ cd /usr/local/autotest/tests/firmware_TouchpadMTB
$ DISPLAY=:0 OPTIONS="-m MANUAL" python main.py
$ DISPLAY=:0 OPTIONS="-m ROBOT" python main.py
Observe the proper firmware version and modes in the directory
and log files.

Change-Id: Ib5038f78721808b27ffe8f533bcfda0935942c35
Reviewed-on: https://gerrit.chromium.org/gerrit/40429
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>