Autotest: Catch drone initialization failures.
If the drone fails initialization (i.e. its disk is completely full),
this will cause the scheduler to not start with the other remaining
drones and will put the lab in a failure loop.

This change catches the AutoservRunError and continues on with the
remaining drones.

BUG=chromium-os:36998
TEST=Checked that the scheduler starts properly when there is no problem.

Change-Id: I3863ed75cac5cdbbdc5fe7e9ee2583d93f94c109
Reviewed-on: https://gerrit.chromium.org/gerrit/40472
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>