firmware_TouchpadMTB: a flexible summary format for variable firmware versions
This CL supports a more flexible summary format which allows the
number of firmware versions to be 1, 2, or more. In addition, it
also displays the counts of tests for every firmware versions, since
there is no guarantee that the counts of tests in every validator
for different firmware versions would be exactly the same. Hence,
it is better to print the counts for every firmware versions so that
the readers could observe it if a mis-match is a problem.

BUG=chromium-os:37494
TEST=Run the command with the $log_dir being the directory
containing the *.log files generated from running the tests.
$ python firmware_summary $log_dir

Observe that it displays the summary according to the number of
firmware versions. It also displays the counts of tests for every
firmware.

Change-Id: I5a475ce08ac37d9f3b8903bcb51c8af1712e162e
Reviewed-on: https://gerrit.chromium.org/gerrit/40521
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>