firmware_TouchpadMTB: support the command line option of iteration
Currently, when recording the gestures in the robot mode, an operator
is required to adjust the robot fingers between the one-finger
setting and the two-finger setting. In order to reduce the manual
operation, it is desirable to record all one-finger gestures for
multiple iterations before proceeding to two-finger gestures.

By recording multiple iterations, it is also possible to reduce
the resultant test result directories.

Step 1: Records gestures for a firmware drop.
Step 1.1: Use the robot to record the gestures in the robot mode.
          Typically, at least 3 iterations are recommended.
          $ DISPLAY=:0 OPTIONS="-m ROBOT -i 3" python main.py
          This would generate a test result directory.
Step 1.2: A tester records the gestures in the manual mode.
          Set the iteration option to be the same as in Step 1.1
          $ DISPLAY=:0 OPTIONS="-m MANUAL -i 3" python main.py
          This would generate another test result directory.
Step 2: Repeat step 1 for the other firmware drop.
Step 3: Derive the summary of the test results.
Step 3.1: copy all logs to a tmp directory, say /tmp/logs.
          Note that $LOG_DIR are the two directories generated above.
          $ cp /var/tmp/touchpad_firmware_test/$LOG_DIR/*.log /tmp/logs
Step 3.2: Derive the summary.
          $ python firmware_summary.py /tmp/logs

BUG=chromium-os:35024
TEST=Run the command on a chromebook with or without the robot.
$ cd /usr/local/autotest/tests/firmware_TouchpadMTB
$ DISPLAY=:0 OPTIONS="-i 3" python main.py
Observe that it will guide the user to record 3 rounds for every
gesture variations.

Change-Id: I3adfa2cc2cabd5f4bff78770468324a37dc5f151
Reviewed-on: https://gerrit.chromium.org/gerrit/40522
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>