firmware_TouchpadMTB: fix a bug about the gestures in the robot_sim mode
When the mode is set to be 'robot_sim', i.e. robot simulation mode,
the gestures to record should be the same as those in the 'robot' mode
instead of those in the 'robot_int' mode.

Note that the robot_sim mode would only be used for debugging purpose
by the developer.

BUG=chromium-os:35751
TEST=Run the command on a chromebook without the robot.
$ cd /usr/local/autotest/tests/firmware_TouchpadMTB
$ DISPLAY=:0 OPTIONS="-s -m robot_sim" python main.py
Observe that it guides the user to record the following gestures only:
  - ONE_FINGER_TRACKING
  - ONE_FINGER_TO_EDGE
  - ONE_FINGER_SWIPE
  - ONE_FINGER_TAP
  - ONE_FINGER_PHYSICAL_CLICK
  - TWO_FINGER_TRACKING
  - TWO_FINGER_SWIPE
  - TWO_FINGER_TAP
  - TWO_FINGER_PHYSICAL_CLICK

Change-Id: I0be7aeada1aebe7bdc99c35a56789c609932d3f8
Reviewed-on: https://gerrit.chromium.org/gerrit/40523
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>