Check if the DUT rebooted before ABORTING.
When the DUT reboots after a suspend/resume, processes that have
been scheduled to run, exit, and we loose our ssh connection.
These events happen simultaneously, but autotest, only noticing the
exit code and a malformed last line, throws an unexpected ABORT.
This change attempts to diagnose the failure by first pinging the DUT,
then querying it for a boot_id over ssh before ABORTING.

Note this patch does not change the saved boot_id when the test
requests a reboot. Therefore we can only tell if a reboot happened,
not if it is unexpected.

TEST=Triggered the unexpected ABORT by restarting during suspend/resume.
BUG=chromium-os:37596

Change-Id: I8390f20e047034ccb9585591ccd16cf71bfd7050
Reviewed-on: https://gerrit.chromium.org/gerrit/40628
Commit-Queue: Prashanth Balasubramanian <beeps@chromium.org>
Reviewed-by: Prashanth Balasubramanian <beeps@chromium.org>
Tested-by: Prashanth Balasubramanian <beeps@chromium.org>