Add in graphics_SyncControlTest to autotests
This test checks that sync control extension behave as expected. Currently this
only implemented for EGL, with the GLX implementation being a place holder for
future work. For EGL there is only one function tested,
eglGetsyncValuesCHROMIUM, since the rest of the spec has not yet been defined
and implemented.

This CL add the implementation of the test as a dep in the glbench dep. There
are other standalone binaries in this dep already, so this seems like the most
resonable place for it to be.

BUG=chrome-os-partner:16624
TEST=Tested binary as a standalone on a EGL device, daisy:
      - `emerge-daisy autotest-deps-glbench`
      - Copied synccontroltest from /build/daisy/... to device
      - `export DISPLAY=:0`
      - `synccontroltest`
      - Confirmed no errors reported
     Confirmed that binary built for a GLX device, link:
      - `emerge-link autotest-deps-glbench`
     Built and ran the autotest, on daisy:
      - `emerge-daisy autotest-deps-glbench`
      - `emerge-daisy autotest-tests`
      - `run_remote_tests.sh --use_emerged --remote=$REMOTE
        graphics_SyncControlTest`
     Confirmed that autotest built for a GLX device, link:
      - `emerge-link autotest-deps-glbench`
      - `emerge-link autotest-tests`

Change-Id: Ia0d737cbd939920cf1c2055e3158901f6094be38
Reviewed-on: https://gerrit.chromium.org/gerrit/41290
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>