Test that the ME region is properly locked.
For systems that have an Intel Management Engine, we need to ensure that the
it has been properly locked down prior to leaving the factory. This is an
explicit test for that.

This should be run as part of the factory finalization tests (and before we
give the ODM the final firmware image too, of course).

BUG=chromium-os:35675
BRANCH=all
TEST=factory,manual

In production, it should be run automatically.

You can run it manually from a chroot like this:

  ./run_remote_tests.sh --board=$BOARD --remote=$DUT_IP_ADDR \
                        client/site_tests/firmware_LockedME/control$

To test, I ran it on Link with the current release firmware
(Google_Link.2695.1.133). That firmware is locked, so the test passes.

I then replaced the Link firmware with a developer firmware, which has an
unlocked ME (note: you have to do this externally - you can't change it with
flashrom, that's what "locked" means), rebooted and re-ran. The test
detected that the ME was unlocked and failed.

On Snow, it fails with a warning message because Snow has no ME.

Change-Id: Ib8d20ae408a7c472f70db520e0a6d8a65d9dbc1f
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/41061
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>