security_SMMLocked: determine state of SMM
This test checks the SMM on Intel devices to make sure it has been locked
and that the SMRAM area is unmapped. ARM is ignored, since it doesn't
have SMM. We'll need to add SMRAM offset locations as new chipsets are
added since there isn't a way to discover it dynamically.  For now,
tie SMRAM offset to the Northbridge identifier.

BUG=chromium-os:37779
TEST=link and daisy build, security_SMMLocked passes.

Change-Id: Iac18a49bba4892c02b63373ce2a54c3836723ac5
Signed-off-by: Kees Cook <keescook@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/41074
Reviewed-by: Jim Hebert <jimhebert@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Brian Harring <ferringb@chromium.org>