firmware_TouchpadMTB: check wrong number of Tracking IDs in test_flow
For multi-finger tests, e.g., two_finger_swipe, it is very
unlikely that only one finger is observed in the generated events
if the operator has performed the multi-finger gesture correctly.
In this CL, a warning message is shown in red color on the prompt
window to alert the operator that fewer tracking IDs (fingers)
have been observed, and prompt the operator to press 'd' to discard
this gesture and record again. The operator still has the option to
press SPACE to save this gesture if he/she believes that the gesture
has been performed correctly.

Note that if more tracking IDs (fingers) are observed, there will
be no warning message as this is a generally a hardware/firmware
mistake that we would like to capture.

BUG=chromium-os:35024
TEST=Run the command on a chromebook.
$ cd /usr/local/autotest/tests/firmware_TouchpadMTB
DISPLAY=:0 OPTIONS="-s -m complete" python main.py

For any 2f tests, performing 1f gesture, e.g., 1f swipe. Observe
that it displays a warning message in red color on the prompt window.

Change-Id: Ibd92e6ac4b6e65f271e1ebb28a68c83e7d30faf0
Reviewed-on: https://gerrit.chromium.org/gerrit/41197
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>