Teach FAFT about the index structures depthcharge uses to find the EC binary.
If FAFT tries to find the EC binary by looking just past the device tree and
fails to even find the device tree, it should then attempt to use the index
structure at the front of the RW section that depthcharge uses. Similarly, if
it fails to update the size of the EC binary in the device tree itself, it
should instead update the index structure.

The exception is caught in _find_ecbin_offset itself because that function's
name suggests it will find the ecbin offset regardless of where it needs to
look. The exception from _find_ecbin_size_offset_on_dtb is caught in the
calling function because that name says specifically where the size offset is
that needs to be updated. If it fails, it shouldn't then do something beyond
what it was asked to do and update the index structure. It's up to the caller
to decide that's necessary and to take care of it.

BUG=chrome-os-partner:16680
TEST=Ran the firmware_UpdateECBin FAFT test on Link on depthcharge firmware.
Before this change, the test bombed out because it got an exception while
while trying to look up the device tree in the RW section of the image. After
this change, the test passed.
BRANCH=None

Change-Id: I114442e0241056fa712b9b413f1171dea14fafe6
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/41191
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>