[autotest] Add a suite_dependencies parameter to dynamic suites
In order to reduce the need for duplicate tests control files
which are identical except for one dependency (for instance,
many network3g tests which come in a tmobile and verizon
version), this change adds a feature -- the ability for a suite
to specify suite-level dependencies which are appended to the
test dependencies when tests are run as part of the suite.

The logic for parsing dependencies is elaborate enough that it
made sense to fold it out into its own function, to make it easier
to write a targeted unit test of dependency fetching/parsing.

BUG=chromium-os:37371
TEST=Unit. Also, local run_suite of new suites created in pending change
I52a91e0dfb01712e8ef760d16d122271d59c8b98 creates jobs with correct
dependencies

Change-Id: I53f8ef671f7744a50cad4e07e91885760cb29b61
Reviewed-on: https://gerrit.chromium.org/gerrit/41232
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>