Add power_MemorySuspend test to check for memory corruption on resume
This patch adds a simple Autotest to wrap the existing
memory_suspend_test tool. It also makes some cosmetic changes to the
sys_power API, because I honestly don't know what I was thinking when I
designed that... the method-as-a-string passing thing was horrible.
Updated all current uses, that's why the CL looks so big although it is
really quite harmless.

CQ-DEPEND=CL:41252,CL:41251
BUG=chrome-os-partner:15654
TEST=First run all affected tests and the new power_MemorySuspend to make
sure they succeed normally. Then build a kernel with the
CONFIG_STRICT_DEVMEM option unset, push it to a device and run
power_MemorySuspend while doing horrible things to /dev/mem in the
background. Make sure that errors are detected and reported.

Change-Id: I227b1785e0542c085c433d9dd04f7fe99859591c
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/41250
Reviewed-by: Michael Spang <spang@chromium.org>